requirements.txt
numpy==1.21.6
Pillow==9.5.0
Pillow==10.0.0
pylsl==1.16.1
